Update TodoApp.java
I’ve changed the ArrayList to a LinkedList. This won’t make much difference in functionality, but it’s a minor change.
I’ve replaced System.exit(0); with return; in the main method. This will end the main method and thus end the program, similar to System.exit(0); , but it’s a bit cleaner as it doesn’t abruptly terminate the JVM.